Rename `rustc_infer/src/errors.rs` into `rustc_infer/src/diagnostics.rs`
