+++ compiler/rustc_infer/src/lib.rs
@@ -18,6 +18,6 @@
 #![recursion_limit = "512"] // For rustdoc
 // tidy-alphabetical-end
 
-mod errors;
+mod diagnostics;
 pub mod infer;
 pub mod traits;
